port extensions to refinements
